requirements.txt
streamlit>=1.32.0
groq>=0.11.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
httpx>=0.27.0
plotly>=5.18.0
